+++ src/components/Header.tsx
@@ -1,50 +1,23 @@
-import { useNavigate } from "react-router-dom";
+import { useLocation } from "react-router-dom";
+import SubHeader from "./header/SubHeader";
+import MainHeader from "./header/MainHeader";
 
 type HeaderProps = {
-  title: string;
+  title?: string;
   showPreviousIcon?: boolean;
   showShareIcon?: boolean;
 };
 
 const Header = ({
-  title,
+  title = "",
   showPreviousIcon = true,
   showShareIcon = true
 }: HeaderProps) => {
-  const navigate = useNavigate();
+  const pathname = useLocation().pathname;
+  const isLoggedIn = localStorage.getItem("isLoggedIn");
+  
 
-  const handleGoBack = () => {
-    navigate(-1);
-  };
-
-  return (
-    <div className="relative flex h-[56px] w-full flex-row items-center justify-center border-b border-gray-100 bg-white">
-      {showPreviousIcon && (
-        <img
-          src="/public/icon/arrow_left.svg"
-          alt="Go To Back"
-          className="absolute left-[18.77px] cursor-pointer"
-          width={9.87}
-          height={16}
-          onClick={handleGoBack}
-        />
-      )}
-
-      <h1 className="absolute left-1/2 -translate-x-1/2 transform text-[18px] font-bold text-gray-900">
-        {title}
-      </h1>
-
-      {showShareIcon && (
-        <img
-          src="/public/icon/share.svg"
-          alt="Share"
-          className="absolute right-[20px] cursor-pointer"
-          width={16}
-          height={16}
-        />
-      )}
-    </div>
-  );
+  return pathname === "/" ? <MainHeader isLoggedIn={isLoggedIn}/> : <SubHeader title={title} showPreviousIcon={showPreviousIcon} showShareIcon={showShareIcon} />;
 };
 
 export default Header;
